Conflicts:
	Include/object.h

+++ Include/lock_linux.h
@@ -55,9 +55,9 @@ Py_LOCAL_INLINE(void) futex_unlock_primitive(int *lock) {
 	}
 }
 
-PyAPI_FUNC(long) Py_AtomicInc(long* value);
-PyAPI_FUNC(long) Py_AtomicDec(long* value);
-PyAPI_FUNC(long) Py_AtomicAdd(long* to, long value);
+PyAPI_FUNC(Py_ssize_t) Py_AtomicInc(Py_ssize_t* value);
+PyAPI_FUNC(Py_ssize_t) Py_AtomicDec(Py_ssize_t* value);
+PyAPI_FUNC(Py_ssize_t) Py_AtomicAdd(Py_ssize_t* to, Py_ssize_t value);
 
 
 #endif
